Move D3 dependency into main library
The demo used D3 to setup a blank SVG for the library to draw on. This
is ok in the webpack-based demo, but makes life harder for users of the
library via a more traditional script tag because they also have to have
D3 available somehow. Since we're bundling D3 inside the library at the
moment, it doesn't make sense to require that duplication.

I initially tried to just rewrite the code that required D3 in vanilla
JS. This had a weird issue inside Dagre where the root SVG container was
ending up different (it was a <g> inside the SVG, not the SVG element
itself). Rather than debug a `this` context issue, (which it looked
like) I've just moved responsibility for initialising the SVG with D3
inside the main drawing method.

To do this more reliably, I've also updated the main drawing method so
that you pass in the container element, rather than it assuming it
already exists under a specific id.