Reference images outside of webpack
Packaging images using `require`s is great for consumers using webpack,
but it's tricky for others because they don't have a way of modifying
the path webpack sets for them. At the expense of a little bit more
config for webpack users, I've made the root path configurable and
switched around all of the webpack config to copy the files explicitly.