Handle replaced statments
With this change, the visualisation tool better handles replaced
statements, i.e. data which contains multiple historical statements.

In short, replaced statements are now removed from the visualisation, so
that you see only the most recent version. This is actually more useful
when working with data that replaces statements, because it allows you
to check that you are replacing the right things.