Reduce opacity of ended relationships
One key use of replacesStatements is to mark ownerships as having ended.
Previously, the visualisation tool would not show these any differently,
meaning it was confusing to understand the diagram. The visual language
is not explicit about how to represent these, but we have followed the
principle of Relevance and so reduced the opacity of ended
relationships.